Refactored AutoComplete to seperate class
