+++ src/Services/AutoComplete/IAutoComplete.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace JuliusSweetland.OptiKey.Services.AutoComplete
+{
+    interface IAutoComplete
+    {
+        IEnumerable<string> GetSuggestions(string root);
+    }
+}
